[Feat#2] Nested singleton handling
