+++ lib/decors.rb
@@ -1,6 +1,6 @@
 require 'decors/decorator_base'
 require 'decors/decorator_definition'
-require 'decors/method_added_listener'
+require 'decors/method_added'
 
 module Decors
     VERSION = '0.1.0'
