+++ lib/decors/decorator_definition.rb
@@ -8,9 +8,14 @@ module Decors
             method_definer = mixin ? :define_method : :define_singleton_method
 
             send(method_definer, decorator_name) do |*params, &blk|
-                ctx = self.singleton_class? ? ObjectSpace.each_object(self).first : self
-                ctx.send(:extend, MethodAddedListener)
-                ctx.declared_decorators << [decorator_class, params, blk]
+                if singleton_class?
+                    ObjectSpace.each_object(self).first.send(:extend, ::Decors::MethodAdded::ForwardToSingletonListener)
+                    extend(::Decors::MethodAdded::SingletonListener)
+                else
+                    extend(::Decors::MethodAdded::StandardListener)
+                end
+
+                declared_decorators << [decorator_class, params, blk]
             end
         end
     end
